+++ library/src/main/kotlin/eu/tib/orkg/prototype/community/adapter/output/jpa/internal/OrganizationEntity.kt
@@ -58,7 +58,6 @@ class OrganizationEntity() {
         Organization(
             id = OrganizationId(id!!),
             name = name,
-            logo = null,
             createdBy = ContributorId(createdBy!!),
             homepage = url,
             observatoryIds = observatories!!.map { ObservatoryId(it.id!!) }.toSet(),