+++ library/src/main/kotlin/eu/tib/orkg/prototype/community/domain/model/Organization.kt
@@ -11,8 +11,6 @@ data class Organization(
 
     var name: String?,
 
-    var logo: String?,
-
     @JsonProperty("created_by")
     val createdBy: ContributorId? = ContributorId.createUnknownContributor(),